Added coffee, less, and sass templates/engines
Should we support variations of an engine @cmather?
In other words, if somebody runs the command
`iron create app-name --css=stylus`
should we convert that to .styl, which is the correct file extension
for stylus, or should we force people to be explicit when passing in flags?:
`iron create app-name --css=styl`

If we convert variations, we would need to convert
- styls to .styl
- sass to .scss
- coffeescript to .coffee
- typescript to .ts

What do you think?