Is there any way you can help me out with this real quick @cmather?
I'm still trying to familiarize myself with some of this
code and I'm looking for the right place to honor the
`--skip-template-{css,html,js}` flag.

On line 66 in init.js I was able to prevent the style sheet
generator from getting invoked by checking `config.template.css`,
but right above that the home generator is getting invoked which
creates home.css.

Should we even be checking which templates should be generated
in the init.js file, or do you have something else in mind?