itterate through generators
I'm kind of lost here @cmather.

On line 54 in lib/commands/generate.js, instead of
running only one generator, I'm getting the config
file and I'm trying to run an array of generators.

I've tried messing around with fibers to get each
generator to run in a syncronous manner, but everything
I've tried, only the first generator in the array is run.

Is there a good way to do this from the generator command?