roll back changes to command.js and generate.js
Explanation:

In general we want to leave the semantics of top level commands alone
and instead build functionality into the specific generators or new
generators. Or create new top level commands if needed. In this case, we
just want to invoke some additional generators when creating a route. So
the order of invocation in pseudo code would be:

```
run init
  invoke route
    injectAtEndOfFile(...)
    invoke template
    invoke controller
```

Affects:

- c77fc541c6a5e97a5a1975eb67670843cdab9681
- 6e279148aaa35f41b07c652735aa0feba8defa70
- bd5c7c6b3c5ee71cb7d27683f1a324fe20e86818
- ab9342e216c5dcdc101514439f3bf3525a0cfdc4
- 1bce7ac6e84ca370783929da76c3183c1c3c0100