fix 100% cpu when invoking meteor commands
Maybe fixes issue #120

For some reason setting the stdio: 'inerhit' option to the spawn command
when spawning the meteor command was causing a cpu spike. If I instead
use the default which is to open pipes between stdin, stdout and stderr
and add a callback to the stdout data event, the meteor command runs
with normal cpu.